Add TR reload log + factorization
